Support update_urls in upload subscriptions (bug 769)
The server will now send a list of tuples that has
the key "update_urls" in a response to a subscription
upload. The client needs to parse this and return a
new object to the calling code.

Added new unit tests and modified existing tests to
accomodate for the 2010-01-07 API changes.